chore(providers): rename from settings to options
