+++ packages/cerebras/src/cerebras-chat-options.ts
