+++ packages/deepinfra/src/deepinfra-completion-options.ts
@@ -1,4 +1,4 @@
-import { DeepInfraChatModelId } from './deepinfra-chat-settings';
+import { DeepInfraChatModelId } from './deepinfra-chat-options';
 
 // Use the same model IDs as chat
 export type DeepInfraCompletionModelId = DeepInfraChatModelId;
