+++ packages/deepinfra/src/deepinfra-embedding-options.ts
