+++ packages/deepseek/src/deepseek-chat-options.ts
