+++ packages/fireworks/src/fireworks-chat-options.ts
