+++ packages/fireworks/src/fireworks-completion-options.ts
