+++ packages/lmnt/src/lmnt-provider.ts
@@ -1,7 +1,7 @@
 import { SpeechModelV1, ProviderV1 } from '@ai-sdk/provider';
 import { FetchFunction, loadApiKey } from '@ai-sdk/provider-utils';
 import { LMNTSpeechModel } from './lmnt-speech-model';
-import { LMNTSpeechModelId } from './lmnt-speech-settings';
+import { LMNTSpeechModelId } from './lmnt-speech-options';
 
 export interface LMNTProvider extends Pick<ProviderV1, 'speechModel'> {
   (
