+++ packages/lmnt/src/lmnt-speech-options.ts
