+++ packages/openai/src/internal/index.ts
@@ -9,5 +9,5 @@ export * from '../openai-image-settings';
 export * from '../openai-transcription-model';
 export * from '../openai-transcription-settings';
 export * from '../openai-speech-model';
-export * from '../openai-speech-settings';
+export * from '../openai-speech-options';
 export * from '../responses/openai-responses-language-model';
