+++ packages/openai/src/openai-provider.ts
@@ -31,7 +31,7 @@ import { OpenAITranscriptionModelId } from './openai-transcription-settings';
 import { OpenAIResponsesLanguageModel } from './responses/openai-responses-language-model';
 import { OpenAIResponsesModelId } from './responses/openai-responses-settings';
 import { OpenAISpeechModel } from './openai-speech-model';
-import { OpenAISpeechModelId } from './openai-speech-settings';
+import { OpenAISpeechModelId } from './openai-speech-options';
 
 export interface OpenAIProvider extends ProviderV2 {
   (modelId: 'gpt-3.5-turbo-instruct'): OpenAICompletionLanguageModel;
