+++ packages/openai/src/openai-speech-options.ts
