+++ packages/xai/src/xai-chat-options.ts
